[MOV] website_crm_partner_assign: split models and views into separate files
Give a little coup de jeune to this module by reorganizing models and views
into separate files.

Task-2752043
