+++ addons/website_crm_partner_assign/models/__init__.py
@@ -3,4 +3,6 @@
 
 from . import crm_lead
 from . import res_partner
+from . import res_partner_activation
+from . import res_partner_grade
 from . import website
