+++ addons/website_crm_partner_assign/models/res_partner.py
@@ -3,37 +3,6 @@
 
 from odoo import api, fields, models
 from odoo.osv import expression
-from odoo.addons.http_routing.models.ir_http import slug
-
-
-class ResPartnerGrade(models.Model):
-    _name = 'res.partner.grade'
-    _order = 'sequence desc'
-    _inherit = ['website.published.mixin']
-    _description = 'Partner Grade'
-
-    sequence = fields.Integer('Sequence')
-    active = fields.Boolean('Active', default=lambda *args: 1)
-    name = fields.Char('Level Name', translate=True)
-    partner_weight = fields.Integer('Level Weight', default=1,
-        help="Gives the probability to assign a lead to this partner. (0 means no assignment.)")
-
-    def _compute_website_url(self):
-        super(ResPartnerGrade, self)._compute_website_url()
-        for grade in self:
-            grade.website_url = "/partners/grade/%s" % (slug(grade))
-
-    def _default_is_published(self):
-        return True
-
-
-class ResPartnerActivation(models.Model):
-    _name = 'res.partner.activation'
-    _order = 'sequence desc'
-    _description = 'Partner Activation'
-
-    sequence = fields.Integer('Sequence')
-    name = fields.Char('Name', required=True)
 
 
 class ResPartner(models.Model):
