+++ addons/website_crm_partner_assign/models/res_partner_activation.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class ResPartnerActivation(models.Model):
+    _name = 'res.partner.activation'
+    _order = 'sequence desc'
+    _description = 'Partner Activation'
+
+    sequence = fields.Integer('Sequence')
+    name = fields.Char('Name', required=True)
