+++ addons/website_crm_partner_assign/models/res_partner_grade.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+from odoo.addons.http_routing.models.ir_http import slug
+
+
+class ResPartnerGrade(models.Model):
+    _name = 'res.partner.grade'
+    _order = 'sequence desc'
+    _inherit = ['website.published.mixin']
+    _description = 'Partner Grade'
+
+    sequence = fields.Integer('Sequence')
+    active = fields.Boolean('Active', default=lambda *args: 1)
+    name = fields.Char('Level Name', translate=True)
+    partner_weight = fields.Integer('Level Weight', default=1,
+        help="Gives the probability to assign a lead to this partner. (0 means no assignment.)")
+
+    def _compute_website_url(self):
+        super(ResPartnerGrade, self)._compute_website_url()
+        for grade in self:
+            grade.website_url = "/partners/grade/%s" % (slug(grade))
+
+    def _default_is_published(self):
+        return True
